feat(core): re-write the module from scratch
The library now correctly handles all error codes, status codes and even throws an error in case of
a transport-level exception. The export types now include a singleton "verifyEmail()" function which
takes in the token and the email and returns a formatted response. The library also exports a
Verifier object (class?) which can be used with the new keyword and used over and over again.

BREAKING CHANGE: Module's export definitions have been changed